requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
fastapi-amis-admin==0.5.0
fastapi-user-auth==0.5.0
fastapi-scheduler==0.0.12
apscheduler==3.10.1
SQLAlchemy==1.4.41
aiosqlite==0.19.0
aiomysql==0.2.0
SQLModel==0.0.8
pandas==2.0.3
numpy==1.24.3
modbus-tk==1.1.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==1.10.12
pytz==2023.3
paho-mqtt==1.6.1
structlog==23.1.0
pytest==7.4.0
python-multipart==0.0.6
loguru
suds-py3
pymssql
crypto
